requirements.txt
psutil==5.9.4
PyQt6==6.4.0
python_bcrypt==0.3.2
termcolor==2.1.1
